Efficient circle rendering
Although GPUs are not built for rendering circles, the PICA200's proctex functionality provides the hardware functionality to fake one perfectly in 2D.
This commit adds a bit of internal logic to switch between different proctex setups- the existing one, to blend between the texture and the primary color, and the new one, used to generate a circle.
This circle is generated using proctex's ADDSQRT2 function and using GPU_MIRRORED_REPEAT as the clamping operation. Because ADDSQRT2 has the property of emitting a quarter of a circle, mirroring/repeating it can emit a full circle. The alpha lookup table can then be configured to have full alpha within the circle, and no alpha outside of the circle. Multiplying this alpha value with the current alpha in TexEnv discards any fragments not within the circle.
This is far, far superior to the current scenario in which people go around spamming vertices for triangles to fill in a circle.