Properly set TexEnvSrc when drawing primitives
The comments say that calling `C2Di_SetSrc` is not needed, but that is not true. If you use `C2D_DrawImage`, the `TexEnvSrc` remains set to `GPU_TEXTURE0`. It looks like drawing the primitives afterwards works regardless, but that's only if the last texel of the image rendered had nonzero alpha. If the texture used has enough texels with zero alpha, the zero alpha may remain stuck somewhere in the hardware regs and no primitive is drawn afterwards. Easily tested with rendering `RGB5551` texture set to all zeros. Can be "fixed" by calling `C2D_Prepare()`, but that's an overkill, of course.